GitHub Actions for building the ODBC driver on Windows (#2)
## Summary
Created a workflow for building the ODBC driver for Windows.

## Description
The workflow to build the ODBC driver is triggered on push, pull
request, and manually.
To run the workflow manually, on GitHub, go to Actions > ODBC Driver and
click on "Run workflow". Note: this button will only show up if the
workflow file is in the master branch.
<img width="1643" alt="Screenshot 2025-02-07 at 10 07 01 AM"
src="https://github.com/user-attachments/assets/a0d824fc-0477-4111-a256-537632258cad"
/>
When manually triggering the workflow, a user can specify the commit to
pull for the arrow and flightsql-odbc. If it is left blank then the
commits in the
[build_win64.bat](https://github.com/dbt-labs/warpdrive/blob/master/build_win64.bat)
is used.
<img width="344" alt="Screenshot 2025-02-07 at 10 06 20 AM"
src="https://github.com/user-attachments/assets/471cba5c-5ab6-4b8c-a926-19bacb410d24"
/>

While working through the build issues there are steps added to help
debug and should be removed before merging. TODO comments in the
workflow file are included.

---------

Signed-off-by: Guian Gumpac <guian.gumpac@improving.com>
Co-authored-by: Ravjot Brar <83892020+ravjotbrar@users.noreply.github.com>